requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.2
python-multipart==0.0.6
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
boto3
